+++ js/util.js
@@ -1,0 +1,25 @@
+const getRandomPositiveInteger = (min, max) => {
+  const lower = Math.ceil(Math.min(Math.abs(min), Math.abs(max)));
+  const upper = Math.floor(Math.max(Math.abs(min), Math.abs(max)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+const getRandomPositiveFloat = (min, max, digits = 1) => {
+  const lower = Math.min(Math.abs(min), Math.abs(max));
+  const upper = Math.max(Math.abs(min), Math.abs(max));
+  const result = Math.random() * (upper - lower) + lower;
+  return result.toFixed(digits);
+};
+
+const getRandomArrayElement = (elements) => elements[getRandomPositiveInteger(0, elements.length - 1)];
+
+const getRandomArrayUniqElements = (elements, length) =>
+  [...elements.sort( () => .5 - Math.random() )].splice(0, length === 1 ? 1 : length -1);
+
+export {
+  getRandomArrayElement,
+  getRandomArrayUniqElements,
+  getRandomPositiveFloat,
+  getRandomPositiveInteger
+};
